+++ webui/src/components/ClusterButtonsPanel.js
@@ -1,13 +1,17 @@
 // @flow
 import * as React from 'react';
 import { connect } from 'react-redux';
-import { isBootstrapped } from 'src/store/selectors/clusterPage';
+import { ControlsPanel } from '@tarantool.io/ui-kit';
+import {
+  isBootstrapped,
+  isVshardAvailable
+} from 'src/store/selectors/clusterPage';
+
 import AuthToggleButton from 'src/components/AuthToggleButton';
 import FailoverButton from 'src/components/FailoverButton';
 import BootstrapButton from 'src/components/BootstrapButton';
 import ProbeServerModal from 'src/components/ProbeServerModal';
 import ClusterIssuesButton from 'src/components/ClusterIssuesButton';
-import { ControlsPanel } from '@tarantool.io/ui-kit';
 import type { State } from 'src/store/rootReducer';
 
 type ClusterButtonsPanelProps = {
@@ -50,7 +54,7 @@ const mapStateToProps = (state: State) => {
 
   return {
     showFailover: !!(clusterSelf && clusterSelf.configured),
-    showBootstrap: !isBootstrapped(state),
+    showBootstrap: isVshardAvailable(state) && !isBootstrapped(state),
     showToggleAuth: !implements_add_user && !implements_list_users && implements_check_password
   }
 };